+++ cdl/tests/features/signals/fft1d_unit.py
@@ -21,14 +21,14 @@ from cdl.utils.vistools import view_curves
 
 
 def test():
-    """Signal peak dialog test"""
+    """1D FFT unit test."""
     with qt_app_context():
         newparam = new_signal_param(stype=SignalTypes.COSINUS, size=10000)
         s1 = create_signal_from_param(newparam)
         t, y = s1.xydata
         f, s = xy_fft(t, y)
         t2, y2 = xy_ifft(f, s)
-        execenv.print("Comparing original and FFT/IFFT signals:", end=" ")
+        execenv.print("Comparing original and FFT/iFFT signals...", end=" ")
         np.testing.assert_almost_equal(t, t2, decimal=3)
         np.testing.assert_almost_equal(y, y2, decimal=10)
         execenv.print("OK")