+++ cdl/utils/vistools.py
@@ -11,10 +11,15 @@ import guiqwt.image
 import guiqwt.plot
 import guiqwt.tools
 from guidata.configtools import get_icon
+from guidata.qthelpers import win32_fix_title_bar_background
+from guiqwt.baseplot import BasePlot
 from guiqwt.builder import make
+from guiqwt.plot import ImagePlot, PlotManager, SubplotWidget
+from qtpy import QtCore as QC
+from qtpy import QtWidgets as QW
 
 from cdl.config import _
-from cdl.utils.qthelpers import exec_dialog
+from cdl.utils.qthelpers import exec_dialog, qt_app_context
 from cdl.utils.tests import get_default_test_name
 
 TEST_NB = {}
@@ -116,3 +121,87 @@ def view_images(data_or_datalist, name=None, title=None, xlabel=None, ylabel=Non
         item = make.image(data, interpolation="nearest", eliminate_outliers=0.1)
         items.append(item)
     view_image_items(items, name=name, title=title, xlabel=xlabel, ylabel=ylabel)
+
+
+class SyncPlotWindow(QW.QMainWindow):
+    """Window for showing plots, optionally synchronized"""
+
+    def __init__(self, parent=None, title=None, icon=None):
+        super().__init__(parent)
+        win32_fix_title_bar_background(self)
+        self.setWindowTitle(self.__doc__ if title is None else title)
+        if icon is not None:
+            self.setWindowIcon(get_icon(icon) if isinstance(icon, str) else icon)
+        self.manager = PlotManager(None)
+        self.manager.set_main(self)
+        self.subplotwidget = SubplotWidget(self.manager, parent=self)
+        self.setCentralWidget(self.subplotwidget)
+        toolbar = QW.QToolBar(_("Tools"))
+        self.manager.add_toolbar(toolbar, "default")
+        toolbar.setMovable(True)
+        toolbar.setFloatable(True)
+        self.addToolBar(toolbar)
+
+    def add_panels(self):
+        """Add standard panels"""
+        self.subplotwidget.add_standard_panels()
+
+    def rescale_plots(self):
+        """Rescale all plots"""
+        QW.QApplication.instance().processEvents()
+        for plot in self.subplotwidget.plots:
+            plot.do_autoscale()
+
+    def showEvent(self, event):  # pylint: disable=C0103
+        """Reimplement Qt method"""
+        super().showEvent(event)
+        QC.QTimer.singleShot(0, self.rescale_plots)
+
+    def add_plot(self, row, col, plot, sync=False, plot_id=None):
+        """Add plot to window"""
+        if plot_id is None:
+            plot_id = str(len(self.subplotwidget.plots) + 1)
+        self.subplotwidget.add_subplot(plot, row, col, plot_id)
+        if sync and len(self.subplotwidget.plots) > 1:
+            syncaxis = self.manager.synchronize_axis
+            for i_plot in range(len(self.subplotwidget.plots) - 1):
+                syncaxis(BasePlot.X_BOTTOM, [plot_id, f"{i_plot + 1}"])
+                syncaxis(BasePlot.Y_LEFT, [plot_id, f"{i_plot + 1}"])
+
+
+def __compute_grid(num_objects, max_cols=4, fixed_num_rows=None):
+    """Compute number of rows and columns for a grid of images"""
+    num_cols = min(num_objects, max_cols)
+    if fixed_num_rows is not None:
+        num_rows = fixed_num_rows
+        num_cols = (num_objects + num_rows - 1) // num_rows
+    else:
+        num_rows = (num_objects + num_cols - 1) // num_cols
+    return num_rows, num_cols
+
+
+def view_images_side_by_side(
+    images,
+    titles,
+    share_axes=True,
+    rows=None,
+    maximized=False,
+    title=None,
+):
+    """Show sequence of images"""
+    rows, cols = __compute_grid(len(images), fixed_num_rows=rows, max_cols=4)
+    with qt_app_context(exec_loop=True):
+        win = SyncPlotWindow(title=title, icon="datalab.svg")
+        for idx, (data, imtitle) in enumerate(zip(images, titles)):
+            row = idx // cols
+            col = idx % cols
+            plot = ImagePlot(title=imtitle)
+            item = make.image(data, interpolation="nearest", eliminate_outliers=0.1)
+            plot.add_item(item)
+            win.add_plot(row, col, plot, sync=share_axes)
+        win.add_panels()
+        if maximized:
+            win.resize(1200, 800)
+            win.showMaximized()
+        else:
+            win.show()